enhance(infrastructure): embed URL to current commit in HTML
this ensures that there can be _no_ mistake about what source code you
are observing while you're using the application. this is super helpful
when you're uncertain whether to trust your eyeballs about what source
code you are currently experiencing in a deployed run-time. for example,
sometimes it is difficult to _see_ changes that you have deployed, and
if things are known to be broken, you might find yourself wondering whether
the deployment you're testing _really has_ the commits that you think
it does