enhance(infrastructure): use npm in place of yarn
with the introduction of a lock mechanism to npm, there are few (if any)
advantages to using `yarn` over `npm` in 2020. yarn's roadmap has progressively
aimed it toward being a "universal" (cross-language and cross-platform)
package manager, as opposed to being a primitive improvement over npm.
unless these features are important to us, we bear the on-going cost of
requiring an additional tool and dependency in every developers set-up for
no return.

here are a few things happening in this commit:

- the dev dependencies and run-time dependencies are sorted out in [`./package.json`]()
- all references to `yarn` are replaced with `npm`
- a [`./package-lock.json]`() file is generated in place of [`yarn.lock`]()

issue: https://jira.healthsparq.net/browse/SEAM-5518