enhance(infrastructure): implement BeforeAll ensuring the application is available
this commit implements a BeforeAll hook for the acceptance test suite
that:

1. Checks to see if a complete build artifact of the application is avialable
- technically this could even ensure that the build artifact is one that corresponds
to the current commit using https://github.com/cambiahealth/journi-web-app/pull/11
2. If not, it runs a build of the app
3. When a build is complete, it finds an available port to serve the app on
4. It starts a static server to run the app for the tests
5. It stores information about the dynamic port and server instance on the
global context so that they can be torn down in the AfterAll hook.

this is great, because now running the tests is totally self-contained,
no additional npm scripts are needed, and there's no way for a user to
get confused about the external environment needed to run the cukes.

this also has the side-effect of making the whole test suite run faster (somehow)