refactor(styles): integrate [styled-components](https://styled-components.com/docs)
this commit opens a deeper integration with styled-components as a styling
implementation, in contrast with scss. in particular:

- variables and other thematic values are instantiated in [`./src/styles/theme/index.ts`](https://github.com/cambiahealth/journi-web-app/compare/styled-components-integration?expand=1#diff-7205f0249e4d0cb8e9d435a127947091)
- these values are made available to any styled-component via `props.theme`
- some of these are used to instantiate "global" styles, that is, styles
not specifically scoped to a single component. this facilitates some of the
basic typography, for example colors and a rhythmic font-size hierarchy.
- the structure of the theme object is arbitrary and up to us. but i used
[Google's Material UI component library theme API ](https://material-ui.com/customization/default-theme/?expand-path=$.typography) as my guide here,
because I heard our designer mention that she was using Material UI for
her own guidance and inspiration.
- [polished](https://polished.js.org/docs/#fontface) is a useful css utility library with all kinds of run-time
computational bits and bobs. i grabbed it because i wanted to implement a
classic typographical hierarchy using modular steps. you can see it in use
[here](https://github.com/cambiahealth/journi-web-app/compare/styled-components-integration?expand=1#diff-7205f0249e4d0cb8e9d435a127947091R339). you can read more about the idea here: [in Google's design documents, among other places](https://material.io/design/typography/#). but the basic
output should be roughly the same as before. the difference is that now,
when you have to decide what size something should be, instead of setting
an arbitrary pixel value that "looks good", you can just walk up in steps
using the pre-defined scale: 1,2,3,4, and so on. this ensures a consistent
vertical rhythm throughout the visual product surface.

a bunch of SCSS is left in place just to avoid re-writing it all in one
commit. one alternative is to delete it all and break the app for a minute,
but i want it to be up and running and looking normal for the upcoming demo.

following this commit, the way is fully open to develop any scene or component
using styled-components.

you can see an example of using styled-components to style a page component in the [src/routes/error/index.tsx](https://github.com/cambiahealth/journi-web-app/compare/styled-components-integration?expand=1#diff-86dab2a491981aedb07ab72cd0b44012), and a layout component in [src/layouts/standard/index.tsx](https://github.com/cambiahealth/journi-web-app/compare/styled-components-integration?expand=1#diff-1f20d4a280610aca13c649483c25af33)