enhancement(CI): implement pre-push git hook for building and testing the app
there are a few things that are nice to have happen automatically, rather
than being notified by Circle _after_ a failure, or having to remember
to run them manually:

- "build" the app and styleguide (compile the JS and CSS assets as a production bundle)
- lint the source code using tslint
- execute the visual regression tests against the styleguide

this commit implements a git hook that does all of these things prior
to pushing a commit to the remote. no reason to do all of it for every
commit or anything.

in order for this to be most effective, you may want to write one of these files
on your machine:

```
REACT_APP_STORYBOOK_URL=file://${whatever the absolute path to this repo on your machine is}/styleguide
```

this file is git ignored. the value you input will be used to run the
visual regression tests against a compiled instance of the styleguide.
you can set it to whatever you want, for example: localhost:9009
if you'd prefer to always run the tests against a live running development
server of the styleguide.